requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
google-generativeai==0.3.2
python-dotenv==1.0.0
